Move scan button to add book fragment
