+++ front_end/app/src/main/res/layout/activity_new_item.xml
@@ -35,14 +35,7 @@
             android:textColor="@color/white"
             android:textSize="18dp"
             android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/scan_barcode_button"
-            style="@style/Widget.AppCompat.Button"
-            android:layout_width="50dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Scan" />
+        
     </LinearLayout>
     <LinearLayout
         android:layout_width="match_parent"
